[MCP Extract] Reorganize process server files into folder structure
